move password traits to credential ns
