+++ src/User/Entity/Credential/Features/PasswordProtected.php
@@ -4,14 +4,25 @@ declare(strict_types=1);
 
 namespace MsgPhp\User\Entity\Credential\Features;
 
-use MsgPhp\User\Password\PasswordProtectedTrait;
+use MsgPhp\User\Password\PasswordAlgorithm;
 
 /**
  * @author Roland Franssen <franssen.roland@gmail.com>
  */
 trait PasswordProtected
 {
-    use PasswordProtectedTrait;
+    /** @var string */
+    private $password;
+
+    public function getPassword(): string
+    {
+        return $this->password;
+    }
+
+    public function createPasswordAlgorithm(): PasswordAlgorithm
+    {
+        return PasswordAlgorithm::create();
+    }
 
     abstract public function withPassword(string $password): self;
 }
