+++ src/User/Entity/Credential/Features/PasswordWithSaltProtected.php
@@ -4,16 +4,27 @@ declare(strict_types=1);
 
 namespace MsgPhp\User\Entity\Credential\Features;
 
-use MsgPhp\User\Password\PasswordWithSaltProtectedTrait;
+use MsgPhp\User\Password\{PasswordAlgorithm, PasswordSalt};
 
 /**
  * @author Roland Franssen <franssen.roland@gmail.com>
  */
 trait PasswordWithSaltProtected
 {
-    use PasswordWithSaltProtectedTrait;
+    use PasswordProtected;
 
-    abstract public function withPassword(string $password): self;
+    /** @var string */
+    private $passwordSalt;
+
+    public function getPasswordSalt(): string
+    {
+        return $this->passwordSalt;
+    }
+
+    public function createPasswordAlgorithm(): PasswordAlgorithm
+    {
+        return PasswordAlgorithm::createLegacySalted(new PasswordSalt($this->passwordSalt));
+    }
 
     abstract public function withPasswordSalt(string $passwordSalt): self;
 }
